+++ src/tests.rs
@@ -1,0 +1,5 @@
+use crate::cmd_runner::CommandRunner;
+
+pub fn _get_runner() -> CommandRunner {
+    CommandRunner
+}
